chore(release): prepare v0.8.36
- Bump workspace version to 0.8.36
- Formalize CHANGELOG for v0.8.36
- Bump npm wrapper version and deepseekBinaryVersion
- Sync internal crate dependency pins
- Refresh Cargo.lock
