+++ crates/cli/Cargo.toml
@@ -14,13 +14,13 @@ path = "src/main.rs"
 anyhow.workspace = true
 clap.workspace = true
 clap_complete.workspace = true
-deepseek-agent = { path = "../agent", version = "0.8.35" }
-deepseek-app-server = { path = "../app-server", version = "0.8.35" }
-deepseek-config = { path = "../config", version = "0.8.35" }
-deepseek-execpolicy = { path = "../execpolicy", version = "0.8.35" }
-deepseek-mcp = { path = "../mcp", version = "0.8.35" }
-deepseek-secrets = { path = "../secrets", version = "0.8.35" }
-deepseek-state = { path = "../state", version = "0.8.35" }
+deepseek-agent = { path = "../agent", version = "0.8.36" }
+deepseek-app-server = { path = "../app-server", version = "0.8.36" }
+deepseek-config = { path = "../config", version = "0.8.36" }
+deepseek-execpolicy = { path = "../execpolicy", version = "0.8.36" }
+deepseek-mcp = { path = "../mcp", version = "0.8.36" }
+deepseek-secrets = { path = "../secrets", version = "0.8.36" }
+deepseek-state = { path = "../state", version = "0.8.36" }
 chrono.workspace = true
 dirs.workspace = true
 serde.workspace = true
