+++ crates/core/Cargo.toml
@@ -9,13 +9,13 @@ description = "Core runtime boundaries for DeepSeek workspace architecture"
 [dependencies]
 anyhow.workspace = true
 chrono.workspace = true
-deepseek-agent = { path = "../agent", version = "0.8.35" }
-deepseek-config = { path = "../config", version = "0.8.35" }
-deepseek-execpolicy = { path = "../execpolicy", version = "0.8.35" }
-deepseek-hooks = { path = "../hooks", version = "0.8.35" }
-deepseek-mcp = { path = "../mcp", version = "0.8.35" }
-deepseek-protocol = { path = "../protocol", version = "0.8.35" }
-deepseek-state = { path = "../state", version = "0.8.35" }
-deepseek-tools = { path = "../tools", version = "0.8.35" }
+deepseek-agent = { path = "../agent", version = "0.8.36" }
+deepseek-config = { path = "../config", version = "0.8.36" }
+deepseek-execpolicy = { path = "../execpolicy", version = "0.8.36" }
+deepseek-hooks = { path = "../hooks", version = "0.8.36" }
+deepseek-mcp = { path = "../mcp", version = "0.8.36" }
+deepseek-protocol = { path = "../protocol", version = "0.8.36" }
+deepseek-state = { path = "../state", version = "0.8.36" }
+deepseek-tools = { path = "../tools", version = "0.8.36" }
 serde_json.workspace = true
 uuid.workspace = true
