+++ crates/secrets/src/lib.rs
@@ -202,19 +202,25 @@ impl InMemoryKeyringStore {
 
 impl KeyringStore for InMemoryKeyringStore {
     fn get(&self, key: &str) -> Result<Option<String>, SecretsError> {
-        Ok(self.entries.lock().unwrap().get(key).cloned())
+        let guard = self.entries.lock().map_err(|e| {
+            SecretsError::Keyring(format!("InMemoryKeyringStore mutex poisoned: {e}"))
+        })?;
+        Ok(guard.get(key).cloned())
     }
 
     fn set(&self, key: &str, value: &str) -> Result<(), SecretsError> {
-        self.entries
-            .lock()
-            .unwrap()
-            .insert(key.to_string(), value.to_string());
+        let mut guard = self.entries.lock().map_err(|e| {
+            SecretsError::Keyring(format!("InMemoryKeyringStore mutex poisoned: {e}"))
+        })?;
+        guard.insert(key.to_string(), value.to_string());
         Ok(())
     }
 
     fn delete(&self, key: &str) -> Result<(), SecretsError> {
-        self.entries.lock().unwrap().remove(key);
+        let mut guard = self.entries.lock().map_err(|e| {
+            SecretsError::Keyring(format!("InMemoryKeyringStore mutex poisoned: {e}"))
+        })?;
+        guard.remove(key);
         Ok(())
     }
