+++ crates/tui/Cargo.toml
@@ -21,8 +21,8 @@ path = "src/main.rs"
 [dependencies]
 anyhow = "1.0.100"
 arboard = "3.4"
-deepseek-secrets = { path = "../secrets", version = "0.8.35" }
-deepseek-tools = { path = "../tools", version = "0.8.35" }
+deepseek-secrets = { path = "../secrets", version = "0.8.36" }
+deepseek-tools = { path = "../tools", version = "0.8.36" }
 schemaui = { version = "0.12.0", default-features = false, optional = true }
 async-stream = "0.3.6"
 async-trait = "0.1"
