+++ crates/tui/src/tui/mod.rs
@@ -17,8 +17,9 @@ pub mod approval;
 pub mod auto_router;
 pub mod backtrack;
 pub mod clipboard;
-mod color_compat;
+pub mod color_compat;
 pub mod command_palette;
+pub mod composer_ui;
 pub mod context_inspector;
 pub mod context_menu;
 pub mod diff_render;
@@ -30,6 +31,7 @@ pub mod file_mention;
 pub mod file_picker;
 pub mod file_picker_relevance;
 pub mod file_tree;
+pub mod footer_ui;
 pub mod format_helpers;
 pub mod frame_rate_limiter;
 pub mod history;
@@ -39,6 +41,7 @@ pub mod live_transcript;
 pub mod markdown_render;
 mod mcp_routing;
 pub mod model_picker;
+pub mod mouse_ui;
 pub mod notifications;
 pub mod onboarding;
 pub mod osc8;
