+++ npm/deepseek-tui/package.json
@@ -1,7 +1,7 @@
 {
   "name": "deepseek-tui",
-  "version": "0.8.35",
-  "deepseekBinaryVersion": "0.8.35",
+  "version": "0.8.36",
+  "deepseekBinaryVersion": "0.8.36",
   "description": "Install and run deepseek and deepseek-tui binaries from GitHub release artifacts.",
   "author": "Hmbown",
   "license": "MIT",
